chore(types): Include phone number id in prepare second factor
This commit modifies the type of the prepare
second factor params to also include the
phone number id which we want to use as second
factor.

This aligns the params with the rest of the
prepare methods and also makes it more
future proof, once we add new second factor
options or we allow users to select which
second factor they want to use.