chore(clerk-js): Consolidate ticket based sign up
We currently have two sign up flows that boil down to using a token to
create an account; instance and organizations invitations.
In our BE, we have consolidated the two to simply send a ticket and
specify the "ticket" verification strategy. Our FE code however, still
treats the two flows differently and tries to differentiate between an
instance invitation token and an organization invitation token.
This commit consolidates the two token-based flows into simply passing a
{ticket: "",strategy:"ticket} payload when creating the sign up.
The "__clerk_invitation_token" query parameter is still supported to
ensure backwards compatibility with any pending invitations that might
have been created before the BE consolidation was deployed.