fix(clerk-js): Invitation based sign up optional fields
Allow users to fill in any optional fields that might be available when
signing up with an invitation, instead of signing them up immediately.
If the instance supports optional fields and a user lands to our
<SignUp/> component from an invitation email, give them a chance to
enter additional information instead of immediately signing them up.
For instances that no additional information is available in the form,
the user will be signed up immediately.