feat(clerk-react): Allow usernames to be optional
With the latest BE code changes, we now support
instances with optional usernames.
This means that during sign up, if the username
attribute is enabled, it doesn't necessarily
mean that it's also required.

This commit modifies the sign up component to
reflect that and show that the username field is
optional if it's enabled but not required.

https://www.notion.so/clerkdev/Support-optional-usernames-in-User-settings-3a03178699a748bcba322e268cd66782