feat(clerk-js,types,backend-core): Expose user.external_id attribute (#182)
* fix(clerk-js): Expose user.externalId field

ExternalId is used during imports.

An example usage is if they had an auth0_id, and need to use that to reference the user, instead of the clerk_id, then they put it there.

* fix(backend-core): Expose user.externalId field

ExternalId is used during imports.

An example usage is if they had an auth0_id, and need to use that to reference the user, instead of the clerk_id, then they put it there.