feat(clerk-js): Organization slugs
Added support for organization slugs. Added slug as a parameter in
clerk.createOrganization().
The Organization.create method should continue to work with a single
string argument for the name. Added a deprecation warning and made it so
we can invoke the function with either the name as a single argument, or
an object.