feat(backend-core): Update organization metadata
Added a new method in the organizations API to update and merge
organization metadata. The method name is updateOrganizationMetadata. It
triggers a PATCH /v1/organizations/:id/metadata request to the Clerk API
and accepts publicMetadata and privateMetadata parameters.
The metadata will be merged with any existing metadata on the
organization.