chore(clerk-js): Enhance buildURL utility
BuildURL can no build a URL safely by using the native URL() constructor. It can also build a secondary path and search URL that lives inside the hash of the main URL.

For example:

https://foo.com/bar?qux=42#/hash-bar?hash-qux=42

This is very useful when we need to navigate across pages or between components in a safe way and avoid string concatenations in URLs.