Safe url hash build (#239)
* chore(clerk-js): Introduce a pathJoin utility

* chore(clerk-js): Enhance buildURL utility

BuildURL can no build a URL safely by using the native URL() constructor. It can also build a secondary path and search URL that lives inside the hash of the main URL.

For example:

https://foo.com/bar?qux=42#/hash-bar?hash-qux=42

This is very useful when we need to navigate across pages or between components in a safe way and avoid string concatenations in URLs.

* fix(clerk-js): Fix navigateToFactorTwo

Navigating to two factor after a sign up transfer flow was broken when the sign in URL contained querystring parameters such as redirect_url or after_sign_up_url.

The culprit was the string concatenation was used to build the final factor-two url as we were appending the #/factor-two string after the query strings.

This PR fixes the issue by leveraging the new, smarter buildURL utility.