fix(clerk-js): Navigate to sign up continue in web3
This fix triggers a component rerender when the user needs to be redirected to "sign-up/continue" page after a web3 flow with missing requirements.

The fix follows the same approach as Clerk.handleRedirectCallback but indicates that ClerkJS routing needs refactoring.

One of the issues is that we need to rerender ClerkJS components when navigating from ClerkJS core. Currently this is only possible when we inject customNavigate from the internal useNavigate() hook.