chore(clerk-js): Introduce useDeepEqualMemo to solve rerendering issues
The appearance is an object. A usual pattern in React is to define the object directly on ClerkProvider itself. That can potentially cause performance issues, since a new object reference will be created every time the parent component rerenders causes our AppearanceContext to recalculate the appearance props and update the theme unnecessarily.