chore(clerk-js): Removed deprecation on setSession
In a previous commit, we introduced `setActive`
which would replace the existing `setSession`. As
such, the `setSession` was marked as deprecated.

This commit removes the deprecation warning on
`setSession`. Given how critical this function is,
we decided to avoid switching to `setActive` yet,
until it's fully and thoroughly tested.
Also, its implementation was reverted so that it
avoids calling `setActive` internally.

https://www.notion.so/clerkdev/Add-ClerkJS-method-to-set-the-active-organization-2f551fafdbd14834a7b5519e87c6f254