fix(clerk-js): Fixed bugs in `setActive` caused by new conditions on session
The `setActive` contains a few new conditions on the
`session` param, given that in this case `session` can
also be undefined (which is something we didn't have
in the previous `setSession`).

However, given that we use the current session if we
don't pass a `session`, a lot of these checks can be
removed.

They were wrong in the first place, because they were
also preventing the execution for `session === null`
so the behaviour was different.

https://www.notion.so/clerkdev/Add-ClerkJS-method-to-set-the-active-organization-2f551fafdbd14834a7b5519e87c6f254