fix(clerk-js): Render SignUp form input errors if missing
This commit fixes a regression introduced in our recent changes for Progressive SignUp. The actual problem was that we filtered out values even with empty string value, for our sign up payload request. This issue, lead the form to not render properly any input error if any of them was required and missing