fix(clerk-js): Render phone number input if applicable during sign up invitation flow
This fixes a bug which came up after we released the required/optional options for email & phones. During an invitation flow and the instance has configured required phone numbers, the user landed in the sign-up form. The phone number input wasn't rendered, as we didn't consider it at all during an invitation flow before the new options, and the user wasn't able to complete the sign-up