fix(clerk-js): Fix navigation when an error occurs during progressive sign up flow
This commit fixes a logical error in our redirect logic during an OAuth flow and progressive sign up. If the external account was verified and were still missing requirements for sign up to complete, we redirected to the sign-up continue url. Although the right thing to do, is to first check if there is an error on the external account verification. Image a scenario where an instance uses sign-up restrictions for emails. The user tries to sign-up using an OAuth account whose email is not listed on the allowlist. We should redirect them back on the sign-up url and preview the error, instead of the sign up continue path