fix(clerk-js): Temporarily ignore handleRedirectCallback error
This error is caused when the host app is using React18
and strictMode is enabled. This useEffects runs twice because
the clerk-react ui components mounts, unmounts and mounts again
so the clerk-js component loses its state because of the custom
unmount callback we're using.
This needs to be solved by tweaking the logic in uiComponents.tsx
or by making handleRedirectCallback idempotent