fix(clerk-js): Render both errors in case of emailOrPhone in sign up form
This fixes a bug in the sign up form in the special case of email or phone. When email & phone are both optional and neither of them is provided the form rendered an error only in the input of the active identifier. We change this behavior to also render the respective error in the 'inactive' identifier input as well.

https://www.notion.so/clerkdev/Return-error-for-missing-phone-in-email-or-phone-case-where-no-phone-is-submitted-69492036ccf542cda8e6d020113abe2b