fix(clerk-js): Proceed with token flow if email is the only required field
The reason for this is that SignUp.create needs to be called to know the email of the invitee. If it is called with an
email as the only required field, the sign up is completed and the user is created, so we cannon successfully let the user
fill optional fields in this screen.
 # Please enter the commit message for your changes. Lines starting