feat(remix): Avoid breaking changes in getAuth and rootAuthLoader
These helpers used to accept either a Request or a LoaderFunctionArgs object. Some runtimes do not support env variables, in which case we rely on the args.context object to include the required variables. For this reason we removed Request from the signature, forcing the users to pass in the `args` object and not `args.request`.

However, this is a breaking chagne so for now we're going to support both. This commit can safely be reverted when we release the next major version.