fix(backend): Add compat layer around request
Our initial intention was to release a v5 major version, however we will instead release this as a non-major release. In order to avoid breaking changes, the the last commit adds a compatibility layer around request in clerk/backend so that the method signatures remain the same. This commit can safely be reverted before we release v5.