fix(backend): Prevent illegal invocation error
Invoking the global.fetch without binding it first to the globalObject fails in Cloudflare Workers with an "Illegal Invocation" error. To fix the issue, we have to bind fetch to the
globalThis object. A customer reported this issue using @clerk/remix on Cloudflare Pages.

The globalThis object is supported for Node >= 12.0.

For more information, refer to https://github.com/supabase/supabase/issues/4417