fix(clerk-js): Set show/hide password button type to button
Pressing enter on the sign up form will trigger a click on the first button with type=submit within the form.
While this isn't a problem on Chrome, on Safari the show/reveal password button catches the click shows/reveals the password
instead of allowing the form to be submitted.
Setting type=button explicitly fixes the issue.