feat(clerk-js): Correct mounted component/modal behaviour
* feat(clerk-js): Redirect to home when an unavailable component is mounted

Redirect to the home url whenever a component is mounted without its criteria being met. For example, mounting the UserProfile component without a user present, will result in a redirect and a warning in the console if in development. When opening a modal for a component which should/can not be opened, print a warning in the console if in development.

* feat(clerk-js): Don't render unavailable modals and show warnings

* fix(clerk-js): Only show redirect warning on development

* test(clerk-js): Add redirectToHome tests

* refactor(clerk-js): Dry withRedirectToHome HOCs

* refactor(clerk-js): Refactor clerk requirements for withRedirectToHome HOCs

* chore(clerk-js): Rename clerkRequirements to componentGuards