feat(remix): Add support for custom responses and headers in rootAuthLoader
We used to accept only plain objects as valid return values from the Remix rootAuthLoader helper. This PR adds support for custom instances of the Response class and custom headers.

It's also important to note with this change, the API of the root loader for Remix apps using Clerk is now aligned to the api of any other loader, as both can return the same type of responses.

One limitation that remains is that if a custom non-redirect response is returned, its body must be a valid json object or a valid json string; otherwise we won't be able to inject the Clerk state.