feat(clerk-sdk-node): Introduce generics in request|response of withAuth
This change will help us make the clerk-sdk-node independent of node
frameworks since we moved from express types to http.IncomingMessage
and http.ServerResponse.
This will make the clerk-sdk-node available for all node frameworks
that extend those types.
To avoid breaking changes we will set the default value of generics
to be the express.Request and express.Response but this will change
in v5 version.