fix(clerk-js): Parse all search params in hash-based router
The hash-based router takes into account the fragment part of the url - that is the part after the first #.

All native browser APIs ignore the fragment part because the spec states that this contains user-defined data. As a result, trying to read the search params using `new URL().searchParams` for a URL that contains search params in its fragment won't work.

In this commit we update the hash-based router to read all search params, defined in the URL or its fragment, so all redirect options passed by the user are respected.