fix(backend): Support the Vercel edge-light runtime key
The edge-light runtime key was recently added as a condition name for the edge runtime in this PR: https://github.com/vercel/next.js/pull/45188

However, this PR also introduced a change (which might not be intentional): the `node` runtime key is now a valid condition key for the edge runtime, so it matches before our `default` import.

This PR adds the `edge-light` key before `node` so that we guarantee that the correct imports will always be used.