feat(clerk-js): Add a new flowStartPath property to the router
This addition will enable as to defined custom flow start points via the <Route /> component. This way, we will return to this
path upon calling navigateToFlowStart. If no flowStart is defined, the flowStartPath will be the same as the startPath.

fix(clerk-js): Fix routermock for tests

refactor(clerk-js): Refactor pathFromFullPath