feat(clerk-js,localizations,types): Allow passing additional OAuth scopes in UserProfile component
This commit, enables users to be able to pass during runtime, additional scopes for each OAuth provider in the UserProfile component. This can be useful in the scenario where a user has already connected an account, and the application requires more scopes for its functionality.
In this case, the component will understand that more scopes are required and will prompt the user to reauthorize their account to grand all the new necessary scopes