feat(clerk-js): Add retry mechanism and error muting to all GET requests to FAPI
This commit adds the exponential backoff retry mechanism on all FAPI requests that have the GET method. It also mutes the network errors returning from all the GET requests when the browser is not online or is not an accepted browser.

In this commit all the resources can handle null data object in their fromJSON method.