feat(shared,clerk-js): Report all FAPI network errors only for valid online browsers (#912)
* feat(shared): Add a maxDelay property on runWithExponentialBackOff utility

The maxDelay property adds a cap to the delay added after every try. This property defaults to 0 (no max delay).

* feat(clerk-js): Add retry mechanism and error muting to all GET requests to FAPI

This commit adds the exponential backoff retry mechanism on all FAPI requests that have the GET method. It also mutes the network errors returning from all the GET requests when the browser is not online or is not an accepted browser.

In this commit all the resources can handle null data object in their fromJSON method.

* fix(clerk-js): Cleanup

---------

Co-authored-by: Nikos Douvlis <nikosdouvlis@gmail.com>