fix(clerk-js): Don't take into account auth instance settings in SignUpContinue form
Clerk offers a progressive sign up flow, which means that using an OAuth provider to sign up, users will be able to fill any required and missing fields, such as name, username, in order to complete the flow. There was a bug in our components, where an instance had only Social Connection enabled as first factor, method to sign in, and required other fields (first/last name) which are not used for sign in of course. If a user tried to sign up with an OAuth provider, which doesn't provide these info (first/last name), the sign up wasn't completed but a user end up in blank screen without a chance to fill the missing fields. This was because we were rendering the form conditionally, based in the instance auth factor settings, which were not applicable in this case