fix(clerk-js): Prevent a switch loop when browser autofills again after switching
It seems that some times (it's inconsistent), the browser will autofill the field in SignInStart after the user has chosen to
switch the identifier. If the autofilled value is a phone number, the user will be unable to switch to "email or username",
since the component will detect the "+" value and automatically switch to the phone input once again. This commit makes it so
the auto-switch can only happen the first time. We might be able to solve this in the future by changing the id of the input
depending on the type, but that may cause more problems than it solves.

 # Please enter the commit message for your changes. Lines starting