fix(clerk-js): Update the height of a form field feedback message on mount
This resolves the followin UI bug. When an error message is mounted and only requires one line for the message to be seen and then we quickly change the password to be empty. We need to update the height of the container in order for a longer message to be displayed now.

+ Speed up the mount and unmount animations for the field texts
+ An error message should have priority over the direction
+ Localize success message for good password
+ Remove green border when Password is successful. This commit also removes the relevant code the passed down the `isSuccessful` prop to the Input component