fix(clerk-js): Redirect for FAPI initiated flows if user is signed in
For flows like OAuth2 IDP in dev instances that use URL based session syncing,
we need to add the dev session jwt as a query paramter so FAPI can recognize
it on redirect (i.e. redirect to /oauth/authorize).
Unless the referrer is the sign in url and the user is not signed in,
we perform the redirect with this added query param.