fix(clerk-js): Load clerk with `isInterstitial`
+ Clerk.load now accepts isIntestitial
This is later used for intestitial specific code that needs to run only when that criteria is met.

In this case we want to either replace a query param with another one or simply remove it completely.

This removes the `__clerk_synced` query param for apps in development instances with CSR