fix(clerk-js,nextjs): Append Dev Browser to redirect URL when using middleware and Hosted Pages
NextJs middleware has no access to browser’s local storage and redirects the user before loading the Clerk-js script. That is, the Dev Browser JWT is not getting propagated to Clerk Hosted Pages, when redirecting there for signing in. If Clerk Hosted Pages have a different auth state on its Dev Browser JWT, then the user ends up in an infinite loop, or cannot access localhost.

To solve this, we need to access the localhost dev browser from the Nextjs middleware (server-side), where Clerk-js has not yet loaded. So, we store the Dev Browser JWT in a __clerk_db_jwt cookie, which can be accessed by Clerk’s authMiddleware and then appended to the URL when redirecting from localhost to Clerk Hosted Pages, so that the Dev Browser JWT from localhost overrides the one in Clerk Hosted Pages. The localhost should be the source of truth for the auth state.