fix(clerk-js): Separate FAPI-initiated redirect flows to requiring user input or not
Previously, for all valid FAPI-initiated redirect flows,
we would require the user to be signed in.
Although this was working properly for oauth, it doesn't
work for magic links or other ticket flows.

This commit updates the logic to separate the allowed
FAPI redirect flows to the ones that require user input,
like the oauth, and the ones that don't, e.g. magic links.