feat(localization): translate deleteOrganization and domainSection objects to Korean (#1630)
* Update ko-KR.ts (Added Missing Account Deletion Description)

In the English translation(en-US.ts), the description 'Delete your account and all its associated data' was present. However, this was missing in the Korean translation(ko-KR.ts). 

This update ensures that the Korean translation also includes the information about deleting all data associated with the account.

This change is essential to provide a clearer understanding to users about the consequences of deleting their account.

* Update ko-KR.ts

korean userProfile > dangerSection > deleteAccountDescription update

* Update ko-KR.ts

fix formatting error

* Update ko-KR.ts (deletePage) translation

* Create pink-suns-fly.md

* Update ko-KR.tsfeat(i18n): translate deleteOrganization and domainSection objects to Korean

---------

Co-authored-by: Nikos Douvlis <nikosdouvlis@gmail.com>