fix(clerk-react,remix,shared): Make `exports.types` the first key in `exports` (#1835)
* fix(clerk-react,remix,shared): Move `exports.types` to the top

The TypeScript docs (https://www.typescriptlang.org/docs/handbook/esm-node.html#packagejson-exports-imports-and-self-referencing) recommends so, but it's also because the "exports" field is order-based.

For example, a scenario where both the "types" and "import" condition could be active, "types" should be first so that it matches and returns a .d.ts file, rather than a .js file from the "import" condition.

* Create dull-gifts-yawn.md

* Update dull-gifts-yawn.md

---------

Co-authored-by: Lennart <lekoarts@gmail.com>