chore(clerk-sdk-node): Drop noisy deprecation of unstable_options in sdk-node (#1858)
The `sdk-node` package is extending the backend `Clerk` with some extra methods.
To do this, we expose a different `Clerk` from the `sdk-node` that returns
the `clerkClient` instance of backend with some extra methods.
To add those extra methods to the return value we currently destruct the
`clerkClient` to a new object and add the extra methods. By destructing the
`clerkClient`, all it's methods and properties are being accessed, causing
the `__unstable_options` deprecation warning to trigger and show the warning
every time in applications using the `sdk-node` package.
To resolve this, we introduced an `ExtendedClerk` type as the return value of
the sdk-node `Clerk` and changed the destructing to
`Object.assign(clerkClient, ...extra...)`.