chore(repo): Completely ignore stdio of all detached spawned processes
If we're spawning a detached process, we must completely ingore the cp's stdio, otherwise the cp will exit when the parent process exits. If we want to log the output of a detached process, we must do it manually by streaming the output to the file and reading the file from the parent process.