chore(repo): Prepend commit hash with a char in prereleaseTemp
We recently found out that `npm` may incorrectly treat a part of the version as a number instead of a string when it only contains numeric characters.

This can cause rare issues when a version only contain numeric characters and it begins with a leading `0`, as this `0` will be dropped when `npm` parses the version as a number.

For example, if we release `4.61.1-staging.0175607` it will end up on `npm` as `4.61.1-staging.175607`

In order to fix this, we simply prepend the verson with a non-numeric character