Chore(repo): Enable clerk@v5 release process (#1837)
* chore(repo): Use main-v4 branch for staging and production releases

Until clerk@v5 is ready for release we will be using
the main-v4 branch for staging and production releases
and every commit and changelog entry SHOULD be committed
in both main-v4 and main branches.
The main branch will be used to produce canary releases.

* chore(repo): Introduce canary releases from main

The main branch will produce canary releases until the
next major version is ready for release. To produce a
staging and production release, open a PR from the main-v4
branch.

* chore(repo): Skip prettier format for integration temp directory