chore(repo): Drop private package eslint-config-custom changeset (#2001)
Adding private packages changesets causes the changeset version to
change the package.json of all packages depending on it to use the
specific version even though the version may be defined as `*`.
This change is causing issues in our integration tests because on
setup the npm tries to retrieve the specific version from npm
and fails (cannot find a private package).

We have tried using `ignore` in changeset/config.json but multiple
Validation errors are thrown from the packages depending on it.