fix(localizations): Fix building of retheme subpath exports (#2245)
Before the localizations subpath exports change
we had a single entrypoint, the `index`. Based
on the CLERK_RETHEME env we were using the
index.retheme.ts or the index.ts as entrypoint
during build. To support that type of conditional
output for all the exported subpaths we need to
either define a callback to replace the *.retheme
files with the *.ts files or conditional choose them
in buiild configuration.
Due to some issues with the tsup `onSuccess` callback
i manually defined all the subpath exports.

We need to be careful in the future and add the new
localization files in both package.json#files, tsup.config.ts
and in subpaths.mjs file to support subpath
exports successfully