feat(clerk-react,clerk-js,shared): Dynamically update props based on deep equality
This change was made in an effort to support dynamic values for appearance, localization as modifying this values usually happens during dev where a fast HMR iteration cycle is required. However, we'd also like to support dynamic values for props like redirectUrl, where the value can be calculated based on data available during runtime, even after the Clerk components mount (eg, based on the results of a fetch request).

In order to avoid extra renders, we use the pre-existing dequal utility to make sure that expensive calculations are avoided.