fix(backend): Revert `{ data, errors }` return value from api client
This change will make all the BackendAPI methods to either return
the value or throw error instead of returning `{ data, errors }`.

We also (temporarily) added the `totalCount` property when we expect
an array of resources to allow passing the `total_count` of paginated
responses from the response body to the return value.