fix(backend,nextjs): Revert `{ data, errors }` return values to v4 format for clerkClient methods and jwt/token helpers (#2633)
* fix(backend): Make signJwt consistent by returning error instead of throwing

* fix(backend): Revert `{ data, errors }` return value from api client

This change will make all the BackendAPI methods to either return
the value or throw error instead of returning `{ data, errors }`.

We also (temporarily) added the `totalCount` property when we expect
an array of resources to allow passing the `total_count` of paginated
responses from the response body to the return value.

* fix(backend): Revert `{ data, errors }` return value from jwt/token helpers

The following helpers were reverted to return their `data` or throw an error.
Internally we will keep the `{ data, errors }` format but added a compatibility
layer to avoid exposing this breaking change:

- `import { verifyToken } from '@clerk/backend'`
- `import { signJwt, hasValidSignature, decodeJwt, verifyJwt } from '@clerk/backend/jwt'`

* fix(nextjs): Apply changes related to adding legacy return value compat layer

* fix(clerk-sdk-node): Update examples related to backend legacy return value compat layer

* chore(repo): Add changeset

* feat(backend): Paginated responses now return { data, totalCount } instead of an array

---------

Co-authored-by: Nikos Douvlis <nikosdouvlis@gmail.com>