fix(clerk-js): Fix FAPI initiated redirect flow for OAuth2 IDP flow with email_link verification (#2677)
For the OAuth2 IDP flow, we should not redirect when the referrer is
the sign up url. This way, the second factor can be completed
after a first factor like email verification link. Previously,
users were being redirected back to FAPI /oauth/authorize
prematurely. This change ensures that users will not be
redirected as such and have the chance to complete their second
factor verification, like phone code.