fix(clerk-js): Make email-link return to application instead of AP (#2727)
* fix(clerk-js): Use signInUrl from context as email-link redirectUrl

The signInUrl used in email-link will be retrieved from the props or
context or repo-level or display config (with the same priority as
listed here).
In applications that the sign-in url is defined the
email-link verification will redirect to the application without passing
through the AP first. This results in fixing the handshake for email-
link flows.
Note: Applications using the signInUrl from the Dashboard settings, will
not support the handshake flow and would require an additional redirect
to FAPI/v1/client/handshake on email-link verification to work.

* fix(clerk-js): Make path-based routing the default in redirects

* fix(clerk-js): Fix type

---------

Co-authored-by: Nikos Douvlis <nikosdouvlis@gmail.com>