fix(clerk-react): Replace semver with custom regex in versionSelector (#2760)
* fix(clerk-react): Replace semver with custom regex in versionSelector

The semver package doesn't provide exports map in its package.json. Vitest is ESM-only and relies more on these things whereas Jest is more forgiving. Instead of tweaking build config for clerk-react and given that we only rely on major() and prerelease() and the bundle-size considerations, we replaced semver with custom regexes

* Create shiny-games-buy.md