fix(shared): Account for legacy frontendApi keys in buildPublishableKey
This helper needs to pick the correct prefix (pk_test or pk_live) depending on the frontendApi key that was passed in. This utility did not take into account legacy frontendApi key formats, such as `clerk.boring.sawfly-91.lcl.dev` resulting in AP treating these instances as prod instances instead. That lead to db_jwt reinitializing, redirect flows not completing etc because of the various `isProduction` checks in the codebase