fix(clerk-js): Hide "Add passkey" button in UP when app is satellite
RP ID for web credentials need to match the domain they are registered in but also the domain where they will be used for authentication.

In satellite app we redirect the user to the primary for signing in. Hence, the user needs to register the passkey in the primary's UserProfile.