fix(*): Switch the compilation target to es2019 for all client-side SDK (#3113)
This enables support for older iOS mobile browsers. The client-side Clerk SDK now support all webkit-based browsers from version iOS 12 and upwards.

This change has minimal impact on the bundle-size and the generated code. From now on, changing the target to a newer version should be considered a breaking change that can only be released after a major version bump.