fix(backend): Fix `getToken` returning `null` when resolving the handshake payload (#3764)
This PR fixes an issue on Remix where getToken() returns no token when redirecting back when sign-in and handshake are triggered.

The issue exists because the getToken() is getting the session token from the request cookie, but when returning from an external sign-in page (e.g. OAuth or Account Portal) on the request there is still the __clerk_handshake cookie, because it's not yet resolved. The issue exists only on production, and also does not happen to nextjs, because middleware has already resolved the handshake.

With this PR we pass the session token that exists inside the handshake payload to be returned when calling getToken().