fix(nextjs): Execute logger.commit() on every middleware invocation
We are using the `withLogger` wrapper to handle the logger.commit on
every time the wrapped function is executed.

In `clerkMiddleware` we are wrapping the whole function with the
`withLogger` instead of the returned nextMiddleware that is executed
in every request causing the `logger.commit()` to be executed once
during the initialization of the `clerkMiddleware()` (on 1st request).
To make debug logs working for `clerkMiddleware` on every request we
need to move the `withLogger` to wrap the returned middleware instead.

In `authMiddleware` we are wrapping the returned nextMiddleware which
makes the debugging logs works without any change.