chore(nextjs): Fix server actions detection
`isServerActionRequest` was still depending on the `next-url` header which no longer is available (removed at `14.1.0` for server actions and in `14.2.2` when requesting RSCs.

This PR fixes the issue and prepares the ground for future work that would require these utilities to work properly.

`pagePath` from __nextGetStaticStore is available since `next@13.5.4` until `next@14.2.5` which is the latest stable release